converter almost final - line breaks to be valiated
