+++ src/main/java/pl/med/clinic/ClinicApplication.java
@@ -2,10 +2,7 @@ package pl.med.clinic;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import pl.med.clinic.files.CSVImport;
-import pl.med.clinic.files.CSVImportScanner;
 
-import java.io.File;
 import java.io.IOException;
 
 @SpringBootApplication
